Merged PR 35945: Fix for date filters on standard grid filter components
- Fix for date filters on standard grids (Renewal is not using the standard methods for filters)
#242682
#242397

There was no bug for the configurations grid however it had the same issue on it. This problem is related to how JS handles dates. Dates are a moment in time and a timestamp is ALWAYS a part of a JS Date. When we run into problems is when we need it to represent a DAY in time. The only way to do that is to use a consistent timestamp for that. The code was already setting 23:59:59 for the defaults so I kept that timestamp as the default. That way if the user chooses the same day for from/to, the comparison on whether the choice is valid will pass.

Related work items: #242690, #243022