Merged PR 36260: Normalize logic for To/From dates in react filter components.
Move common function isQueryValid into utils.js
Set the query.to/from params to use 23:59:59 as base timestamps to avoid timezone offset conversions when getting the dates from them for querystring building
#242397

The net effect of this is that the dates contained in the query object are kept with a 23:59:59 timestamp and when generating the querystring params, From dates will use 00:00:00 and To dates will use 23:59:59. This will also allow for single day queries to work properly.

Related work items: #243022