Merged PR 38384: Fix orderid param in invoice ui-service call
Fix orderid param in invoice ui-service call
We need to also be careful with variable naming so that we are not using global variable names in local functions. i.e. 'data' is the main data object passed into the component but then is used in all of the click handler functions in the grid as the param passed into the function which makes the global variable difficult to reference.

Related work items: #251698