Merged PR 38061: Remove quotes.html param logic for quoteId
This page doesnt appear to support param quoteId. It's possbile the quotePreview Url was changed From once displaying the detail of a quote, To now redirecting to the general quotes grid page. Since the quote is still processing in the background it makes sense that the Url is now pointing to the quotes grid page and now does not require the quoteId

*Unable to test this locally

Fixes:
7) Observe that user is redirected to Quotes page and there is "?quoteId=null" at the end of URL

Related work items: #250361