Merged PR 40327: Revert "Merged PR 40312: Add SearchResults POC component to ui.react.frontend"
Revert "Merged PR 40312: Add SearchResults POC component to ui.react.frontend"

This reverts commit e41e24038c97218fe99439d286717220c2009582.

This is breaking the build. We *think* we need to modify the build to perform an auth with the ADO NPM Repo before running npm install but we don't want to hold up other work while we figure that out.

Related work items: #288199