Merged PR 46489: [332621] Style issue on Create Quote Widget
- Fix style issue that seems to be caused by FontAwesome. This change mitigates the problem but it did not originate in our codebase.

These styles have not been touched in nearly 6 months. I am thinking that a change to the styling in font-awesome has caused the height of the hidden svg element to be much larger than previously making each item in the list take up nearly 400px in height.

Related work items: #332630