Merged PR 46627: [333234] - Unable to select SPA on Quote Preview
Fix null object issue with End user.
The problem is that the quote preview is returning an array for EndUser that contains a null object member. The null checks for the end user object are only checking if the array itself is null or contains 0 elements but it contains 1 null element. I don't know if that in itself should be another bug for the ui-service however to mitigate it, I added another layer of null checking on the object before trying to add properties to it.
#333234

Related work items: #333338