Merged PR 50216: [331583] Renewals grid: Fix reset behavior
This change adjusts the search menu down 3px to match the UX design showing the bottom input border. Also adds a spacer so that the mouse doesn't loose focus on the menu and allows it to stays open for user to select an item. This is a follow up to a previous refactor of this search component

Related work items: #351813