Merged PR 50295: [338107] - Megamenu - New Megamenu Structures to build a hierarchical menu structure as v2
Alter Megamenu Structures to build a hierarchical menu structure.
The JS/CSS was changed only enough to get the secondary menus working. Additional work is needed to get the 3rd and 4th levels working.  The JS may need to be completely reworked in order to work properly.

Related work items: #343176, #352836, #354812