Merged PR 61226:  [419194] Renewal details: Fix cancel display and red placeholder
Fixes these three items

cancel works when an existing ship-to is on the quote, but fails to display correctly when there is no option already selected. In both cases it does not call the UI service (Functional but not a biggie unless going to prod)

when cancel without a ship-to already selected shows the SHIP-TO ADDRESS label on the read-only view and should not show this labelnot showing asterisk (Functional but not a biggie unless going to prod)

when valid=false, mandatory=true. Selecting an item shows a red label above the box ß make it like the other labels (minor/med issue)

Related work items: #442374