Merged PR 61876: XS✔ ◾ [442161] - Teaser Hero6 style for APJ
Hero6 style for APJ. This style takes the Title entered by the author and instead of displaying it as white text overlaying the image, will display in a white box with .8 opacity in the upper corner of the teaser.  This needs to be tested with other use cases than the one I have in the image below.

Result:

![image.png](https://dev.azure.com/techdatacorp/0743fde1-7465-47b2-8226-0c3b84142d6e/_apis/git/repositories/4abf2fec-2538-4f04-919b-91dfaf8f5cdd/pullRequests/61876/attachments/image.png)

Related work items: #448736